8.14 Bugfix

+++ src/controllers/videoController.js
@@ -16,18 +16,31 @@ export const watch = async (req, res) => {
 
 export const getEdit = async (req, res) => {
   const { id } = req.params;
+  const {
+    user:{_id},
+  }=req.session;
   const video = await Video.findById(id);
   if (!video) {
     return res.status(404).render("404", { pageTitle: "Video not found." });
   }
+  if (String(video.owner) !== String(_id)){
+    return res.status(403).redirect("/")
+  }
   return res.render("edit", { pageTitle: `Edit: ${video.title}`, video });
 };
+
 export const postEdit = async (req, res) => {
+  const {
+    user:{_id},
+  } = req.session
   const { id } = req.params;
   const { title, description, hashtags } = req.body;
   const video = await Video.exists({ _id: id });
   if (!video) {
     return res.status(404).render("404", { pageTitle: "Video not found." });
+  }
+  if (String(video.owner) !== String(_id)){
+    return res.status(403).redirect("/");
   }
   await Video.findByIdAndUpdate(id, {
     title,
@@ -36,6 +49,7 @@ export const postEdit = async (req, res) => {
   });
   return res.redirect(`/videos/${id}`);
 };
+
 export const getUpload = (req, res) => {
   return res.render("upload", { pageTitle: "Upload Video" });
 };
@@ -68,9 +82,20 @@ export const postUpload = async (req, res) => {
 };
 export const deleteVideo = async (req, res) => {
   const { id } = req.params;
+  const{
+    user : {_id}
+  }= req.session;
+  const video = await Video.findById(id);
+  if(!video){
+    return res.status(404).render("404", {pageTitle: "Video now found"})
+  }
+  if (String(video.owner) !== String(_id) ){
+    return res.status(403).redirect("/")
+  }
   await Video.findByIdAndDelete(id);
   return res.redirect("/");
 };
+
 export const search = async (req, res) => {
   const { keyword } = req.query;
   let videos = [];
